+++ lib/pcanvas_bitmap.dart
@@ -1,5 +1,5 @@
 /// Portable canvas library (in-memory Bitmap implementation).
-library pcanvas.bitmap;
+library;
 
 export 'pcanvas.dart';
 export 'src/pcanvas_bitmap.dart';